:llama:

+++ HLE/Strings/RegexPool.cs
@@ -109,7 +109,7 @@ public sealed partial class RegexPool : IEquatable<RegexPool>, IEnumerable<Regex
     {
         uint patternHash = SimpleStringHasher.Hash(pattern);
         int hash = HashCode.Combine(patternHash, (int)options, timeout);
-        int index = (int)((uint)hash % _buckets.Length);
+        int index = (int)((uint)hash % (uint)_buckets.Length);
         return _buckets[index];
     }
